adding legacy mode for litespi We can choose between the new DDR implementation and the legacy one which is based on SDR cores. This allows us to use the more efficient core on devices that support the DDR components and the legacy core where they are not supported. Signed-off-by: Paweł Sagan <psagan@antmicro.com>
