+++ litespi/clkgen.py
@@ -8,7 +8,29 @@ from migen import *
 
 from litex.soc.integration.doc import AutoDoc, ModuleDoc
 
-from litex.build.io import DDROutput
+from litex.build.io import SDROutput, DDROutput
+
+class DDRLiteSPIClkGen(Module, AutoDoc, ModuleDoc):
+    """SPI Clock generator
+
+    The ``DDRLiteSPIClkGen`` class provides a generic SPI clock generator.
+
+    The class can be combined with DDR PHY Core.
+
+    Parameters
+    ----------
+    pads : Object
+        SPI pads description.
+
+    Attributes
+    ----------
+    en : Signal(), in
+        Clock enable input, output clock will be generated if set to 1, 0 resets the core.
+    """
+    def __init__(self, pads):
+        self.en         = en         = Signal()
+
+        self.specials += DDROutput(i1=en, i2=0, o=pads.clk)
 
 
 class LiteSPIClkGen(Module, AutoDoc, ModuleDoc):
@@ -59,11 +81,47 @@ class LiteSPIClkGen(Module, AutoDoc, ModuleDoc):
         Controls generation of the ``update`` signal.
     """
     def __init__(self, pads, device, cnt_width=8, with_ddr=False):
+        self.div        = div        = Signal(cnt_width)
+        self.sample_cnt = sample_cnt = Signal(cnt_width)
+        self.update_cnt = update_cnt = Signal(cnt_width)
+        self.posedge    = posedge    = Signal()
+        self.negedge    = negedge    = Signal()
+        self.sample     = sample     = Signal()
+        self.update     = update     = Signal()
         self.en         = en         = Signal()
-        self.active     = active     = Signal()
+        cnt             = Signal(cnt_width)
         en_int          = Signal()
+        clk             = Signal()
 
-        self.comb += self.active.eq(en | en_int)
+        self.comb += [
+            posedge.eq(en & ~clk & (cnt == div)),
+            negedge.eq(en & clk & (cnt == div)),
+            sample.eq(cnt == sample_cnt),
+            update.eq(cnt == update_cnt),
+        ]
+
+        # Delayed edge to account for IO register delays.
+        self.posedge_reg  = posedge_reg  = Signal()
+        self.posedge_reg2 = posedge_reg2 = Signal()
+
+        self.sync += [
+            posedge_reg.eq(posedge),
+            posedge_reg2.eq(posedge_reg),
+        ]
+
+        self.sync += [
+            If(en | en_int,
+                If(cnt < div,
+                    cnt.eq(cnt+1),
+                ).Else(
+                    cnt.eq(0),
+                    clk.eq(~clk),
+                )
+            ).Else(
+                clk.eq(0),
+                cnt.eq(0),
+            )
+        ]
 
         if not hasattr(pads, "clk"):
             # Clock output needs to be registered like an SDROutput.
@@ -94,5 +152,4 @@ class LiteSPIClkGen(Module, AutoDoc, ModuleDoc):
             else:
                 raise NotImplementedError
         else:
-            self.specials += DDROutput(i1=active, i2=0, o=pads.clk)
-
+            self.specials += SDROutput(i=clk, o=pads.clk)
