+++ litespi/phy/generic.py
@@ -9,188 +9,11 @@ from migen.genlib.cdc import MultiReg
 from migen.genlib.misc import WaitTimer
 
 from litespi.common import *
-from litespi.clkgen import LiteSPIClkGen
 
-from litex.soc.interconnect import stream
 from litex.soc.interconnect.csr import *
 
-from litex.build.io import DDRTristate
 
 from litex.soc.integration.doc import AutoDoc, ModuleDoc
-
-# Output enable masks for the tri-state buffers, data mode mask is not included as oe pins default to 0
-
-# LiteSPI PHY Core ---------------------------------------------------------------------------------
-
-class LiteSPIPHYCore(Module, AutoCSR, AutoDoc, ModuleDoc):
-    """LiteSPI PHY instantiator
-
-    The ``LiteSPIPHYCore`` class provides a generic PHY that can be connected to the ``LiteSPICore``.
-
-    It supports single/dual/quad/octal output reads from the flash chips.
-
-    The following diagram shows how each clock configuration option relates to outputs and input sampling in DDR mode:
-
-    .. wavedrom:: ../../doc/ddr-timing-diagram.json
-
-    Parameters
-    ----------
-    pads : Object
-        SPI pads description.
-
-    flash : SpiNorFlashModule
-        SpiNorFlashModule configuration object.
-
-    device : str
-        Device type for use by the ``LiteSPIClkGen``.
-
-    Attributes
-    ----------
-    source : Endpoint(spi_phy2core_layout), out
-        Data stream.
-
-    sink : Endpoint(spi_core2phy_layout), in
-        Control stream.
-
-    cs : Signal(), in
-        Flash CS signal.
-    """
-    def __init__(self, pads, flash, device, clock_domain, cs_delay):
-        self.source              = source = stream.Endpoint(spi_phy2core_layout)
-        self.sink                = sink   = stream.Endpoint(spi_core2phy_layout)
-        self.cs                  = Signal()
-
-        if hasattr(pads, "miso"):
-            bus_width = 1
-            pads.dq   = [pads.mosi, pads.miso]
-        else:
-            bus_width = len(pads.dq)
-
-        assert bus_width in [1, 2, 4, 8]
-
-        # Check if number of pads matches configured mode.
-        assert flash.check_bus_width(bus_width)
-
-        self.addr_bits  = addr_bits  = flash.addr_bits
-        self.ddr        = ddr        = flash.ddr
-
-        assert not ddr
-
-        # Clock Generator.
-        self.submodules.clkgen = clkgen = LiteSPIClkGen(pads, device, with_ddr=ddr)
-
-        # CS control.
-        cs_timer = WaitTimer(cs_delay + 1) # Ensure cs_delay cycles between XFers.
-        cs_out   = Signal()
-        self.submodules += cs_timer
-        self.comb += cs_timer.wait.eq(self.cs)
-        self.comb += cs_out.eq(cs_timer.done)
-        self.comb += pads.cs_n.eq(~cs_out)
-
-        # I/Os.
-        data_bits = 32
-
-        dq_o1  = Signal(len(pads.dq))
-        dq_o2  = Signal(len(pads.dq))
-        dq_i1  = Signal(len(pads.dq))
-        dq_i2  = Signal(len(pads.dq))
-        dq_oe1 = Signal(len(pads.dq))
-        dq_oe2 = Signal(len(pads.dq))
-
-        for i in range(len(pads.dq)):
-            self.specials += DDRTristate(
-                io  = pads.dq[i],
-                o1  = dq_o1[i],
-                o2  = dq_o2[i],
-                oe1 = dq_oe1[i],
-                oe2 = dq_oe2[i],
-                i1  = dq_i1[i],
-                i2  = dq_i2[i]
-            )
-
-        # FSM.
-        shift_cnt = Signal(8, reset_less=True)
-        addr      = Signal(addr_bits if not ddr else addr_bits + addr_width, reset_less=True)
-        data      = Signal(data_bits, reset_less=True)
-
-        usr_dout  = Signal(len(sink.data),  reset_less=True)
-        usr_din   = Signal(len(sink.data),  reset_less=True)
-        usr_len   = Signal(len(sink.len),   reset_less=True)
-        usr_width = Signal(len(sink.width), reset_less=True)
-        usr_mask  = Signal(len(sink.mask),  reset_less=True)
-
-        clk_en = Signal()
-
-        self.comb += self.clkgen.en.eq(clk_en)
-
-        din_width_cases = {1: [
-                 NextValue(usr_din, Cat(dq_o1[1], usr_din)),
-            ]
-        }
-        for i in [2, 4, 8]:
-            din_width_cases[i] = [
-                NextValue(usr_din, Cat(dq_o1[0:i], usr_din))
-            ]
-
-        dout_width_cases = {}
-        for i in [1, 2, 4, 8]:
-            dout_width_cases[i] = [
-                dq_i2.eq(usr_dout[-i:]),
-                NextValue(dq_i1, dq_i2)
-            ]
-
-        self.submodules.fsm = fsm = FSM(reset_state="IDLE")
-        fsm.act("IDLE",
-            sink.ready.eq(cs_out),
-            NextValue(clk_en, 0),
-            If(sink.valid & sink.ready,
-                Case(sink.cmd, {
-                    USER: [
-                        NextValue(usr_dout,  sink.data << (32-sink.len)),
-                        NextValue(usr_din,   0),
-                        NextValue(usr_len,   sink.len),
-                        NextValue(usr_width, sink.width),
-                        NextValue(usr_mask,  sink.mask),
-                        Case(sink.width, dout_width_cases),
-                        NextState("USER")
-                    ]
-                })
-            )
-        )
-
-        fsm.act("USER",
-            NextValue(clk_en, 1),
-            dq_oe2.eq(usr_mask),
-            NextValue(dq_oe1, dq_oe2),
-
-            Case(usr_width, dout_width_cases),
-            Case(usr_width, din_width_cases),
-            NextValue(usr_dout, usr_dout<<usr_width),
-
-            NextValue(shift_cnt, shift_cnt+usr_width),
-            If(shift_cnt == (usr_len-usr_width),
-                NextValue(shift_cnt, 0),
-                NextState("USER_END"),
-            ),
-        )
-        fsm.act("USER_END",
-            NextValue(clk_en, 0),
-            Case(usr_width, din_width_cases),
-            NextValue(dq_i1, 0),
-            NextValue(shift_cnt, shift_cnt+usr_width),
-            If(shift_cnt == (2*usr_width),
-                NextValue(shift_cnt, 0),
-                NextState("SEND_USER_DATA"),
-            ),
-        )
-        fsm.act("SEND_USER_DATA",
-            source.valid.eq(1),
-            source.last.eq(1),
-            source.data.eq(usr_din),
-            If(source.ready,
-                NextState("IDLE"),
-            )
-        )
 
 # LiteSPI PHY --------------------------------------------------------------------------------------
 
@@ -214,6 +37,9 @@ class LiteSPIPHY(Module,AutoDoc, AutoCSR,  ModuleDoc):
     clock_domain : str
         Name of LiteSPI clock domain.
 
+    default_divisor : int (only legacy mode)
+        Default frequency divisor for clkgen.
+
     Attributes
     ----------
     source : Endpoint(spi_phy2core_layout), out
@@ -226,8 +52,14 @@ class LiteSPIPHY(Module,AutoDoc, AutoCSR,  ModuleDoc):
         Flash CS signal from ``LiteSPIPHYCore``.
     """
 
-    def __init__(self, pads, flash, device="xc7", clock_domain="sys", cs_delay=10):
-        self.phy = LiteSPIPHYCore(pads, flash, device, clock_domain, cs_delay)
+    def __init__(self, pads, flash, device="xc7", clock_domain="sys", default_divisor=9, cs_delay=10, legacy=True):
+        if legacy:
+            from litespi.phy.legacy_generic import LiteSPIPHYCore
+            self.phy = LiteSPIPHYCore(pads, flash, device, clock_domain, default_divisor, cs_delay)
+        else:
+            from litespi.phy.ddr_generic import DDRLiteSPIPHYCore
+            self.phy = DDRLiteSPIPHYCore(pads, flash, cs_delay)
+
         self.flash = flash
 
         self.source = self.phy.source
